More configuration and CI fixes
