--- Sources/TMDB/Configuration/BackdropSize.swift
@@ -1,13 +1,0 @@
-public struct BackdropSize: Codable, Hashable, RawRepresentable, Sendable {
-  public var rawValue: String
-
-  public init(rawValue: String) {
-    self.rawValue = rawValue
-  }
-}
-
-extension BackdropSize: ExpressibleByStringLiteral {
-  public init(stringLiteral value: String) {
-    self.rawValue = value
-  }
-}
