--- Sources/TMDB/Configuration/PosterSize.swift
@@ -1,13 +1,0 @@
-public struct PosterSize: Codable, Hashable, RawRepresentable, Sendable {
-  public var rawValue: String
-
-  public init(rawValue: String) {
-    self.rawValue = rawValue
-  }
-}
-
-extension PosterSize: ExpressibleByStringLiteral {
-  public init(stringLiteral value: String) {
-    self.rawValue = value
-  }
-}
