+++ Sources/TMDB/Shared/ImagePath.swift
@@ -1,5 +1,5 @@
 public protocol ImagePath: RawRepresentable<String> {
   associatedtype ImageSize: RawRepresentable<String>
 
-  static var keyPath: KeyPath<Images, [ImageSize]> { get }
+  static var keyPath: KeyPath<ConfigurationDetails.Images, [ImageSize]> { get }
 }
